backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
beautifulsoup4==4.12.3
requests==2.31.0
playwright==1.41.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
stripe==7.11.0
python-dotenv==1.0.0
email-validator==2.3.0
httpx==0.26.0
pyjwt[crypto]==2.8.0
